Move IConsoleDriver to its own file
